pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langchain>=0.3.25",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.5",
    "ollama>=0.4.8",
    "pandas>=2.2.3",
    "protobuf==3.20.*",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "streamlit-chat>=0.1.1",
    "watchdog>=6.0.0",
]
